+++ src/lib.rs
@@ -29,8 +29,6 @@ token positions.
 ## Benefits
 
 - All tokens contain `StrSpan` objects which contain a position of the data in the original document.
-- Supports basic text escaping with `xml:space` (should be invoked manually).
-  A properer text escaping is very hard without the DOM construction.
 - Good error processing. All error types contain position (line:column) where it occurred.
 - No heap allocations.
 - No dependencies.
@@ -64,7 +62,6 @@ token positions.
 mod error;
 mod stream;
 mod strspan;
-mod text;
 mod token;
 mod xml;
 mod xmlchar;
@@ -72,7 +69,6 @@ mod xmlchar;
 
 pub use error::*;
 pub use stream::*;
-pub use text::*;
 pub use strspan::*;
 pub use token::*;
 pub use xml::*;